roominfo_read is unsigned
git-svn-id: file:///Users/walter/linux/walter/src/svnroot/trunk/bbs100@685 d90a1141-1c03-0410-82d0-813645769019

+++ src/Joined.c
@@ -36,7 +36,7 @@ Joined *j;
 	if ((j = (Joined *)Malloc(sizeof(Joined), TYPE_JOINED)) == NULL)
 		return NULL;
 
-	j->roominfo_read = -1;			/* never read it */
+	j->roominfo_read = (unsigned int)-1;		/* never read it */
 	return j;
 }
